# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Discussion Board API"
    },
    "basePath": "/api",
    "schemes": [
      "http"
    ],
    "paths": {
      "/post":{
        "post":{
          "summary":"Create a new post",
          "produces":[
            "application/json"
          ],
          "parameters":[
            {
              "name":"post",
              "in":"body",
              "description":"Post to create",
              "required":true,
              "schema":{
                "$ref":"#/definitions/Post"
              }
            }
          ],
          "responses":{
            "200":{
              "description":"OK",
              "schema":{
                "$ref":"#/definitions/Post"
              }
            },
            "500":{
              "description":"Internal Server Error",
              "schema":{
                "$ref":"#/definitions/Error"
              }
            }
          }
        }
      },
      "/post/comment/{id}":{
        "post":{
          "summary":"Create comment on a post",
          "produces":[
            "application/json"
          ],
          "responses":{
            "200":{
              "description":"OK"
            },
            "500":{
              "description":"Internal Server Error",
              "schema":{
                "$ref":"#/definitions/Error"
              }
            }
          }
        }
      },
      "/post/comments":{
        "get":{
          "summary":"Get all comments on a post",
          "produces":[
            "application/json"
          ],
          "responses":{
            "200":{
              "description":"OK",
              "schema":{
                "type":"array",
                "items":{
                  "$ref":"#/definitions/Comment"
                }
              }
            },
            "500":{
              "description":"Internal Server Error",
              "schema":{
                "$ref":"#/definitions/Error"
              }
            }
          }
        }
      },
      "/user/{id}": {
        "get": {
          "description": "Get a user by ID",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the user to get",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "$ref": "#/definitions/User"
              }
            },
            "404": {
              "description": "User not found",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/user/posts/{id}": {
        "get": {
          "produces": [
            "application/json"
          ],
          "summary": "Get all posts from a user",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the user to retrieve posts for",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Post"
                }
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
       "/posts": {
      "get": {

        "summary": "Get all posts",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Post"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
    },
    "definitions": {
      "Comment":{
        "type":"object",
        "properties":{
          "id":{
            "type":"string"
          },
          "post_id":{
            "type":"string"
          },
          "author_id":{
            "type":"string"
          },
          "content":{
            "type":"string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "author_id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
  